fix(op-devstack): drop misplaced kona tests and finalize componentid cleanup
