+++ op-devstack/stack/component_id.go
@@ -2,6 +2,7 @@ package stack
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"log/slog"
 
@@ -14,6 +15,25 @@ import (
 type ComponentKind string
 
 var _ slog.LogValuer = (*ComponentKind)(nil)
+
+// ChainIDProvider presents a type that provides a relevant ChainID.
+type ChainIDProvider interface {
+	ChainID() eth.ChainID
+}
+
+// KindProvider presents a type that provides a relevant ComponentKind. E.g. KindL2Batcher.
+type KindProvider interface {
+	Kind() ComponentKind
+}
+
+// Keyed presents a type that provides a relevant string key. E.g. a named superchain.
+type Keyed interface {
+	Key() string
+}
+
+const maxIDLength = 100
+
+var errInvalidID = errors.New("invalid ID")
 
 func (k ComponentKind) LogValue() slog.Value {
 	return slog.StringValue(string(k))
