+++ rust/kona/tests/node/common/conductor_test.go
@@ -54,7 +54,7 @@ func TestConductorLeadershipTransfer(gt *testing.T) {
 
 		idToConductor := make(map[string]conductorWithInfo)
 		for _, conductor := range conductors {
-			conductorId := strings.TrimPrefix(conductor.String(), stack.ConductorKind.String()+"-")
+			conductorId := strings.TrimPrefix(conductor.String(), stack.KindConductor.String()+"-")
 			idToConductor[conductorId] = conductorWithInfo{conductor, consensus.ServerInfo{}}
 		}
 		for _, memberInfo := range membership.Servers {
