+++ rust/kona/tests/node/utils/mixed_preset_with_conductor.go
@@ -12,7 +12,7 @@ import (
 type MinimalWithConductors struct {
 	*MixedOpKonaPreset
 
-	ConductorSets map[stack.L2NetworkID]dsl.ConductorSet
+	ConductorSets map[stack.ComponentID]dsl.ConductorSet
 }
 
 func NewMixedOpKonaWithConductors(t devtest.T) *MinimalWithConductors {
@@ -20,7 +20,7 @@ func NewMixedOpKonaWithConductors(t devtest.T) *MinimalWithConductors {
 	orch := presets.Orchestrator()
 	orch.Hydrate(system)
 	chains := system.L2Networks()
-	conductorSets := make(map[stack.L2NetworkID]dsl.ConductorSet)
+	conductorSets := make(map[stack.ComponentID]dsl.ConductorSet)
 	for _, chain := range chains {
 		chainMatcher := match.L2ChainById(chain.ID())
 		l2 := system.L2Network(match.Assume(t, chainMatcher))
